+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # ps (development version)
 
+* New function `errno()` returns a table of `errno.h` error codes and
+  their description.
+
 # ps 1.3.2
 
 * ps now compiles again on unsupported platforms like Solaris.